requirements.txt
flask
certifi==2019.11.28
chardet==3.0.4
geojson==2.5.0
idna==2.8
Pillow==7.0.0
pyowm==2.10.0
python-dateutil==2.8.1
requests==2.22.0
six==1.14.0
suntime==1.2.5
urllib3==1.25.8
spidev
RPi.GPIO
